+++ src/main/java/org/csanchez/jenkins/plugins/kubernetes/PodTemplateUtils.java
@@ -186,6 +186,7 @@ public class PodTemplateUtils {
 
         Map<String, String> podAnnotations = mergeMaps(parent.getMetadata().getAnnotations(),
                 template.getMetadata().getAnnotations());
+        Map<String, String> podLabels = mergeMaps(parent.getMetadata().getLabels(), template.getMetadata().getLabels());
 
         Set<LocalObjectReference> imagePullSecrets = new LinkedHashSet<>();
         imagePullSecrets.addAll(parent.getSpec().getImagePullSecrets());
@@ -217,7 +218,7 @@ public class PodTemplateUtils {
 //        toolLocationNodeProperties.addAll(template.getNodeProperties());
 
         MetadataNested<PodBuilder> metadataBuilder = new PodBuilder().withNewMetadataLike(parent.getMetadata()) //
-                .withAnnotations(podAnnotations);
+                .withAnnotations(podAnnotations).withLabels(podLabels);
         if (!Strings.isNullOrEmpty(template.getMetadata().getName())) {
             metadataBuilder.withName(template.getMetadata().getName());
         }
@@ -257,6 +258,9 @@ public class PodTemplateUtils {
         if (parent == null) {
             return template;
         }
+
+        LOGGER.log(Level.FINEST, "Combining pod templates, parent: {0}", parent);
+        LOGGER.log(Level.FINEST, "Combining pod templates, template: {0}", template);
 
         String name = template.getName();
         String label = template.getLabel();
@@ -306,8 +310,31 @@ public class PodTemplateUtils {
         podTemplate.setAnnotations(new ArrayList<>(podAnnotations));
         podTemplate.setNodeProperties(toolLocationNodeProperties);
         podTemplate.setNodeUsageMode(nodeUsageMode);
+        podTemplate.setInheritFrom(!Strings.isNullOrEmpty(template.getInheritFrom()) ? 
+                                   template.getInheritFrom() : parent.getInheritFrom());
+        
+        podTemplate.setInstanceCap(template.getInstanceCap() != Integer.MAX_VALUE ? 
+                                   template.getInstanceCap() : parent.getInstanceCap());
+        
+        podTemplate.setSlaveConnectTimeout(template.getSlaveConnectTimeout() != PodTemplate.DEFAULT_SLAVE_JENKINS_CONNECTION_TIMEOUT ? 
+                                           template.getSlaveConnectTimeout() : parent.getSlaveConnectTimeout());
+
+        podTemplate.setIdleMinutes(template.getIdleMinutes() != 0 ? 
+                                   template.getIdleMinutes() : parent.getIdleMinutes()); 
+
+        podTemplate.setActiveDeadlineSeconds(template.getActiveDeadlineSeconds() != 0 ?
+                                             template.getActiveDeadlineSeconds() : parent.getActiveDeadlineSeconds()); 
+
+            
+        podTemplate.setServiceAccount(!Strings.isNullOrEmpty(template.getServiceAccount()) ? 
+                                      template.getServiceAccount() : parent.getServiceAccount());
+
+        podTemplate.setCustomWorkspaceVolumeEnabled(template.isCustomWorkspaceVolumeEnabled() ? 
+                                                    template.isCustomWorkspaceVolumeEnabled() : parent.isCustomWorkspaceVolumeEnabled());
+
         podTemplate.setYaml(template.getYaml() == null ? parent.getYaml() : template.getYaml());
 
+        LOGGER.log(Level.FINEST, "Pod templates combined: {0}", podTemplate);
         return podTemplate;
     }
 
@@ -345,7 +372,9 @@ public class PodTemplateUtils {
                     parent = combine(parent, unwrap(next, allTemplates));
                 }
             }
-            return combine(parent, template);
+            PodTemplate combined = combine(parent, template);
+            LOGGER.log(Level.FINEST, "Combined parent + template is {0}", combined);
+            return combined;
         }
     }